Add SelectedTags property to postEditRequest
This update introduces a new `SelectedTags` property to the `postEditRequest` object. It is populated with an array of tag IDs converted to strings from the `post.Tags` collection. This enhancement allows for easier access to selected tag IDs, complementing the existing `Tags` property that provides a list of `SelectListItem` objects.